Finally added in the credits
Some of the codes from the BBCode website did not translate the same - namely justify to have the text be in the middle. I also connected the Credits button to the Credits.tscn. Next lesson I will addd a button to go back to Menu.tscn and resize the rich-text-box so it fits the screen properly.